requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
Django==4.1.3
django-autoslug==1.9.8
django-js-asset==2.0.0
django-mptt==0.14.0
django-stubs==1.13.0
django-stubs-ext==0.7.0
greenlet==2.0.1
msgpack==1.0.4
mypy==0.991
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
pynvim==0.4.3
pytils==0.4.1
sqlparse==0.4.3
tomli==2.0.1
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
